push

+++ cpp/cpp06/ex01/include/Serializer.hpp
@@ -1,7 +1,7 @@
 #ifndef SERIALIZER_HPP
 #define SERIALIZER_HPP
 
-#include <cstdint>
+#include <cstddef>
 #include "Data.hpp"
 
 class Serializer 
@@ -13,8 +13,8 @@ class Serializer
         ~Serializer();
 
     public:
-        static uintptr_t serialize(Data* ptr);
-        static Data* deserialize(uintptr_t raw);
+        static unsigned long serialize(Data* ptr);
+        static Data* deserialize(unsigned long raw);
 };
 
 #endif
