+++ cpp/cpp06/ex01/src/Serializer.cpp
@@ -13,10 +13,12 @@ Serializer& Serializer::operator=(const Serializer& other)
 
 Serializer::~Serializer() {}
 
-uintptr_t Serializer::serialize(Data* ptr) {
-    return reinterpret_cast<uintptr_t>(ptr);
+unsigned long  Serializer::serialize(Data* ptr) 
+{
+    return reinterpret_cast<unsigned long>(ptr);
 }
 
-Data* Serializer::deserialize(uintptr_t raw) {
+Data* Serializer::deserialize(unsigned long raw)
+{
     return reinterpret_cast<Data*>(raw);
 }
